+++ CHANGELOG.md
@@ -4,10 +4,10 @@ A preview of major changes can be found in the Wiki ([Latest Changes](https://gi
 
 ## [unreleased] - TBD
 #### Feature
+- Custom announcer for scores, match finishes, and more!
 - Added new game type `Kill Bull`
 - Added support for stream of board for each player using external stream URLs
 - `Elo Changelog` page for player
-- Custom announcer voice for scores `100` to `180`
 
 #### Changed
 - Added `Outshot Type` and `Tic-Tac-Toe` board to Spectate page